Unmount history component when hidden
- history service resets as history component is unmounted
- restores functionality I changed in my previous commit
